Additional changes to get the DynamoDB Streams working.  Again, this relies on the very latest kappa code which is available only on a branch in the kappa repo.
